Shelter additional constraints

+++ app/controllers/shelters_controller.rb
@@ -11,8 +11,13 @@ class SheltersController < ApplicationController
   end
 
   def create
-    Shelter.create(shelter_params)
-    redirect_to '/shelters'
+    shelter = Shelter.new(shelter_params)
+    if !shelter.save
+      flash[:failure] = "Oh no! Please fill out missing fields before hitting the submit button."
+      redirect_to '/shelters/new'
+    else
+      redirect_to '/shelters'
+    end
   end
 
   def edit
@@ -21,13 +26,25 @@ class SheltersController < ApplicationController
 
   def update
     shelter = Shelter.find(params[:id])
-    shelter.update(shelter_params)
-    redirect_to "/shelters/#{shelter.id}"
+    shelter.assign_attributes(shelter_params)
+    if !shelter.save
+      flash[:failure] = "Oh no! Please fill out missing fields before hitting the submit button."
+      redirect_to "/shelters/#{shelter.id}/edit"
+    else
+      redirect_to "/shelters/#{shelter.id}"
+    end
   end
+
+  # create = new + save
+  # update = asssign_attributes + save
 
   def destroy
     shelter = Shelter.find(params[:id])
-    shelter.destroy
+    if shelter.has_all_adoptable_pets?
+      shelter.delete_pets
+      shelter.delete_reviews
+      shelter.destroy
+    end
     redirect_to '/shelters'
   end
